pyproject.toml
[tool.poetry]
name = "resx"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-phonenumbers = "^1.0.1"
django-phonenumber-field = "^8.0.0"
django-filter = "^24.2"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
python-crontab = "^3.2.0"
aiogram = "^3.10.0"
redis = "^5.0.7"
hiredis = "^2.3.2"
requests = "^2.32.3"
coverage = "^7.6.0"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
